--- a/backend/headless/input_device.c
+++ b/backend/headless/input_device.c
@@ -7,6 +7,7 @@
 #include <wlr/interfaces/wlr_touch.h>
 #include <wlr/util/log.h>
 #include "backend/headless.h"
+#include "util/defs.h"
 #include "util/signal.h"
 
 static void input_device_destroy(struct wlr_input_device *wlr_dev) {
@@ -19,10 +20,12 @@ static struct wlr_input_device_impl input_device_impl = {
 	.destroy = input_device_destroy,
 };
 
+WLR_API
 bool wlr_input_device_is_headless(struct wlr_input_device *wlr_dev) {
 	return wlr_dev->impl == &input_device_impl;
 }
 
+WLR_API
 struct wlr_input_device *wlr_headless_add_input_device(
 		struct wlr_backend *wlr_backend, enum wlr_input_device_type type) {
 	struct wlr_headless_backend *backend =