--- a/backend/libinput/backend.c
+++ b/backend/libinput/backend.c
@@ -5,6 +5,7 @@
 #include <wlr/backend/session.h>
 #include <wlr/util/log.h>
 #include "backend/libinput.h"
+#include "util/defs.h"
 #include "util/signal.h"
 
 static int wlr_libinput_open_restricted(const char *path,
@@ -128,6 +129,7 @@ static struct wlr_backend_impl backend_impl = {
 	.destroy = wlr_libinput_backend_destroy
 };
 
+WLR_API
 bool wlr_backend_is_libinput(struct wlr_backend *b) {
 	return b->impl == &backend_impl;
 }
@@ -154,6 +156,7 @@ static void handle_display_destroy(struct wl_listener *listener, void *data) {
 	wlr_libinput_backend_destroy(&backend->backend);
 }
 
+WLR_API
 struct wlr_backend *wlr_libinput_backend_create(struct wl_display *display,
 		struct wlr_session *session) {
 	assert(display && session);
@@ -185,6 +188,7 @@ error_backend:
 	return NULL;
 }
 
+WLR_API
 struct libinput_device *wlr_libinput_get_device_handle(struct wlr_input_device *_dev) {
 	struct wlr_libinput_input_device *dev = (struct wlr_libinput_input_device *)_dev;
 	return dev->handle;