--- a/include/wlr/backend.h
+++ b/include/wlr/backend.h
@@ -23,6 +23,4 @@ void wlr_backend_destroy(struct wlr_backend *backend);
 struct wlr_egl *wlr_backend_get_egl(struct wlr_backend *backend);
 struct wlr_renderer *wlr_backend_get_renderer(struct wlr_backend *backend);
 
-uint32_t usec_to_msec(uint64_t usec);
-
 #endif