… types/wlr_cursor.c
… types/wlr_data_device.c
… types/wlr_idle.c
… types/wlr_input_device.c
… types/wlr_keyboard.c
… types/wlr_list.c
… types/wlr_output.c
… types/wlr_output_damage.c
… types/wlr_output_layout.c
… types/wlr_pointer.c
… types/wlr_primary_selection.c
… types/wlr_region.c
… types/wlr_screenshooter.c
… types/wlr_seat.c
… types/wlr_server_decoration.c
… types/wlr_surface.c
… types/wlr_tablet_pad.c
… types/wlr_tablet_tool.c
… types/wlr_touch.c
… types/wlr_wl_shell.c
… types/wlr_xcursor_manager.c
… types/wlr_xdg_shell.c
… types/wlr_xdg_shell_v6.c
… util/log.c
… util/os-compatibility.c
… util/region.c
… xcursor/wlr_xcursor.c
… xwayland/selection.c
… xwayland/xwayland.c
… xwayland/xwm.c