This explicitly exports API functions, and sets the default visibility to hidden.
It's a pretty typical and recommended thing for libraries to do.

Yes, it requires a GNU extension (appropriately hidden behind a feature test macro), but I believe this is a justified exception.

Compiled with `-Dbuildtype=minsize -Db_lto=true` and opening rootston
With patch:
```
     22109:	
     22109:	runtime linker statistics:
     22109:	  total startup time in dynamic loader: 27169614 cycles
     22109:		    time needed for relocation: 15428710 cycles (56.7%)
     22109:	                 number of relocations: 4717
     22109:	      number of relocations from cache: 2556
     22109:	        number of relative relocations: 16210
     22109:		   time needed to load objects: 10641900 cycles (39.1%)
```
`436K	libwlroots.so.0.0.0`

Without patch:
```
     22987:	
     22987:	runtime linker statistics:
     22987:	  total startup time in dynamic loader: 42604509 cycles
     22987:		    time needed for relocation: 25563909 cycles (60.0%)
     22987:	                 number of relocations: 4763
     22987:	      number of relocations from cache: 2560
     22987:	        number of relative relocations: 16194
     22987:		   time needed to load objects: 15456870 cycles (36.2%)
```
`460K	libwlroots.so.0.0.0`

As you can see, it makes loading libwlroots a fair bit faster, and makes the final binary a little bit smaller.
It may also be faster, because the LTO can do more, but I have not benchmarked this.