@SirCmpwn 2018-02-19
On my 3.5 GHz CPU, using your numbers, this would save us about 0.001 seconds. I've said no to this before and I'm going to say no to it again.