@ascent12 2018-02-19
Part of it is also not accidentally exporting functions or running into symbol collisions.
If a library user defines a function with the same name as one of our internal functions, they may run into linker issues.

Another example is os_create_anonymous_file. Our examples are reaching into the library internals, and using a function that we don't have in our headers.
That's just asking for trouble.

This creates a much more explicit and clean separation for the library. There is a good reason that like 99% of libraries do this.