@SirCmpwn 2018-02-19
Propose it for standardization. Until then, no.