@ascent12 2018-02-19
It's part of the ELF standard.