@ascent12 2018-02-19
Do you have an actual technical reason for rejecting this?
I'm not going to accept shit without justification.

Yes, non-standard shit isn't good, but it's hidden behind a feature test macro, and we've already have a __GNUC__ test in log.h since basically the very beginning.