@SirCmpwn 2018-02-19
I don't use nonstandard features. That test in log.h really bothers me, I shouldn't have agreed to it.