@ascent12 2018-02-19
The benefits outweigh the consequences. The only tangible downside is a "bad feeling".
There are no effects on the semantics of the code, so any compiler not compatible with GCC will handle it just fine.

The log.h one gives us extra warnings, which prevents a whole bunch of pointless bugs.
This one removes the ability for people to fuck with library internals, and allows an LTO to work better.