@SirCmpwn 2018-02-19
No, the tangible downside is encourange the use of and proliferation of nonstandard features, which I am staunchly against and always have been. You know how this PR was going to go down. Why did you even bother? Just to create animosity?