@ascent12 2018-02-19
I think we need to remove the pretence that our code is portable and standards conforming. There is all sorts of VERY Linux-specific shit we're doing, but you're not complaining because instead of being behind \_\_GNUC\_\_, they're being disabled by \_\_FreeBSD\_\_. I just don't think this "absolutely portable" argument holds any weight.

We're even using other non-standard shit like ## comma deletion and empty variadic macros. Maybe it'll be more relevant when -Wpedantic comes up clean.

I'm not saying that we should go around making everything non-portable; our effort to make things more portable made it even possible for this to run on FreeBSD. I just think there are certain exceptions which are justified, because they provide extremely useful features. Especially things as benign as this and __attribute__((printf)).