@SirCmpwn 2018-02-19
There are several kinds of portability. There's OS portability, libc portability, and compiler portability, to name a few important ones. The nature of this beast is that we aren't gonna get OS portability simply because the whole point is to leverage OS-specific features. But I'll be damned if we're gonna give up on libc or compiler portability, too.

>We're even using other non-standard shit like ## comma deletion and empty variadic macros. Maybe it'll be more relevant when -Wpedantic comes up clean.

Let's file tickets and get this fixed.