@ascent12 2018-02-19
That is going to break probably 50% of our calls to wlr_log, and make it way harder to use.