@SirCmpwn 2018-02-19
We don't need to change wlr_log right away. The value-add of that one is much higher than of ELF visibility, so it's easier for me to look the other way.