@ascent12 2018-02-19
>But I'll be damned if we're gonna give up on libc or compiler portability, too

That's why feature test macros exist.

>The value-add of that one is much higher than of ELF visibility, so it's easier for me to look the other way.

Sure, this isn't the most significant of improvements, but preventing
https://github.com/swaywm/wlroots/blob/master/examples/screenshot.c#L129
is more than enough reason for me.

Funnily enough, you can't even hide ## behind macros, but you can for this. It's quite the double standard.