@SirCmpwn 2018-02-19
I was willing to consider dropping warnings for wlr_log, but we need to use ## for the filename and line number which is a really damn compelling feature. If we're going to let it slide a little, then we might as well do this too.

But let the record state that I think this sucks.