@agx 2018-02-19
What about using a symbols file here instead:

ftp://ftp.gnu.org/old-gnu/Manuals/ld-2.9.1/html_node/ld_25.html

or in actual use:

https://github.com/libvirt/libvirt/blob/master/src/libvirt_admin_private.syms
https://github.com/libvirt/libvirt/blob/master/src/libvirt_admin_public.syms

This prevents cluttering the C file and is supported by GNUs ld as well as lld.

https://releases.llvm.org/3.9.0/tools/lld/docs/ReleaseNotes.html#symbol-versioning

We could simply make wlr_* public and everything else private and at a later point introduce actual versions if needed.